small fixes

+++ src/player.h
@@ -21,7 +21,7 @@ class Player
     m.second += score;
   }
 
-  void finalize(std::vector<std::tuple<int, double>>& global_matchups)
+  void finalize(std::vector<std::tuple<size_t, size_t>>& global_matchups)
   {
     std::for_each(accum_matchups_.begin(), accum_matchups_.end(), [this](auto& scores)
     {
@@ -68,8 +68,8 @@ class Player
   double total_score_ = 0;
   double played_ = 0;
 
-  absl::flat_hash_map<int, std::pair<int, double>> accum_matchups_;
-  std::vector<std::tuple<int, double>> matchups_;
+  absl::flat_hash_map<size_t, std::pair<size_t, double>> accum_matchups_;
+  std::vector<std::tuple<size_t, size_t>> matchups_;
 
   int opponent_start_ = 0;
   int opponents_ = 0;
