+++ src/ratings.h
@@ -22,15 +22,15 @@ class RatingsCalc
 
   private:
 
-  //<index, played vs, score>
-  using Opponent = std::tuple<int, double>;
+  //<index, played vs>
+  using Opponent = std::tuple<size_t, size_t>;
 
   absl::flat_hash_map<std::string_view, int> players_;
   absl::flat_hash_map<int, std::string_view> player_names_;
   std::vector<Player> player_info_;
   std::vector<Opponent> opponent_info_;
   std::vector<double> errors_;
-  std::vector<double> played_;
+  std::vector<size_t> played_;
   int next_player_ = 0;
   int games_ = 0;
 
@@ -45,7 +45,7 @@ class RatingsCalc
   void process_line(std::string_view line);
   double calculate_errors();
   void adjust_ratings_driver(int i, float e);
-  void calculate_errors(int start, int end);
+  void calculate_errors(size_t start, size_t end);
   void adjust_ratings(size_t start, size_t end);
   std::vector<ThreadPool::ThreadJob> create_error_calculation();
   std::vector<ThreadPool::ThreadJob> create_adjust_calculation();
